dynamic argument loading refactor - file rename
