+++ elrond-wasm/src/io/dyn_arg_input_cd.rs
