+++ elrond-wasm/src/io/mod.rs
@@ -1,18 +1,18 @@
 pub mod arg_de_input;
-pub mod arg_loader_endpoint;
-pub mod arg_loader_cd;
-pub mod arg_loader;
 pub mod arg_serialize;
+pub mod dyn_arg_input_endpoint;
+pub mod dyn_arg_input_cd;
+pub mod dyn_arg_input;
 pub mod dyn_arg;
 pub mod macro_helpers;
 pub mod finish;
 pub mod signal_error;
 
 pub use arg_de_input::*;
-pub use arg_loader_endpoint::*;
-pub use arg_loader_cd::*;
-pub use arg_loader::*;
 pub use arg_serialize::*;
+pub use dyn_arg_input_endpoint::*;
+pub use dyn_arg_input_cd::*;
+pub use dyn_arg_input::*;
 pub use dyn_arg::*;
 pub use macro_helpers::*;
 pub use finish::*;
